Separate the config settings for development and automated testing
I added a new configuration setting config.app.development so that
the concerns of development and automated testing could be
separated. These concerns are similar but not the same.  An example is
sub creation, where it's helpful to the developer to not have to level
up their user account, but any automated test of sub creation should
see the same behavior that real users would get.

Setting current_app.config['TESTING'] is already the documented way to
change the behavior of Flask extensions to support automated tests.