Require a password to change the recovery email address
Combined the change password and change email functions onto a single
page, and removed the ability to change emails from the preferences
page.  If validating emails is configured, send a confirmation on
email changes and don't use the new email until it is confirmed.  Hide
the Delete Account option unless on the Edit Account page.