Better handle changes to auth.provider config setting (#120)
After a site adminstrator changes the auth.provider setting from LOCAL
to KEYCLOAK, make Keycloak accounts for pre-existing users whenever
they enter their password, and keep track of their IDs on the Keycloak
server in UserMetadata.

Fix the reset password and change email flows to not create local
passwords or change emails for users with Keycloak accounts if the
Keycloak server is currently not configured, because this would cause
the local account and Keycloak to contain different information if the
Keycloak server is then re-configured (as might happen when a site
admin is learning Keycloak administration).  Add error messages to
tell the users to try again later.

Pin the dependency python-keycloak because they have some current
issues which would require changes to our code if fixed.  See
github.com/marcospereirampj/python-keycloak, issues 90, 95 and 99.