Move server name configuration from app section to site section
Values in the app section are passed through to Flask, which passes
them through to Werkzeug, which complains and refuses to route if you
access a server by its IP address when SERVER_NAME is defined.
If a server is behind a load balancer which terminates https,
it may always be accessed by its IP address.   Move the server name
to the site section of the config so that the server can know the
domain it is serving without confusing Werkzeug.