docker-db: set PGUSER to resolve healthcheck error
POSTGRES_USER is specific to the docker image, and doesn't get picked up
by pg_isready, causing the database to log an error about a missing
'root' role.

Setting PGUSER appropriately causes pg_isready to check the database
with the expected postgres user, allowing it to succeed and eliminating
the error from the logs.