docker: whitelist config.yaml
The config object is created whenever throat.py is invoked,
so docker should include the config.yaml file as standard.

The special case of mounting config.yaml for the migration
service has been removed from the compose file as it's no
longer necessary with the change to the .dockerignore file.