dockerfile: use multi-stage build to help caching
Separating out particularly the npm/webpack stage should help with
caching as the npm buld will only have to be run when the
package*.json, webpack config, or the translation files change.

Switched also to using python:3-slim-buster for the built image,
so there's no need to install python and pip through apt.