dockerfile: 1-char change to appease the linter
This should have no practical effect except to silence the warning from
the Dockerfile linter and remove the problem reported by it in VS Code.

Strictly, the multi-file form of COPY should have a destination ending
in a path separator ( / or \ ), although practically . is a directory
(the root in this case).