docker: don't mask throat dir with bind mount
This commit changes what is mounted into the container with `make up` (or
`docker-compose up`).

Previously, the entire repo was mounted at /throat, but this has the
unfortunate side effect of masking any file created or altered during the
Docker build process that is copied into /throat with the version that exists
on the host (or nothing).

This is the reason, for example, that it was necessary to ensure that
`app/manifest.json` was present on the host before attempting to bring up the
application. It also affects the other files built by webpack and the
translations when they are built via Docker.

This change mounts only app/html and app/templates so these can still be edited
and changes will be picked up by the application as before.

This commit also removes the instruction in the readme to generate the
manifest.json file before developing via Docker, as this is now not necessary
as manifest.json is created when Docker runs `npm run build`.