Change docker config to fix errors & improve caching (#346)
* docker-db: set PGUSER to resolve healthcheck error

POSTGRES_USER is specific to the docker image, and doesn't get picked up
by pg_isready, causing the database to log an error about a missing
'root' role.

Setting PGUSER appropriately causes pg_isready to check the database
with the expected postgres user, allowing it to succeed and eliminating
the error from the logs.

* docker-migration: disable caching

Explicitly disable caching for migration via an environment variable.

Migration 27 fails if the cache is configured but not actually available,
as it makes a call on the config object that otherwise attempts to
fetch from and store to the cache. The cache itself is not needed for
the migration.

* docker: whitelist cli/ in .dockerignore

Commit a6496f77 moved the cli package out from under app, so this
change just ensures it is still included by docker and available
for throat.py to import.

* docker: whitelist config.yaml

The config object is created whenever throat.py is invoked,
so docker should include the config.yaml file as standard.

The special case of mounting config.yaml for the migration
service has been removed from the compose file as it's no
longer necessary with the change to the .dockerignore file.

* docker: ignore python cache files

* dockerfile: Ensure translations are compiled

* dockerfile: use multi-stage build to help caching

Separating out particularly the npm/webpack stage should help with
caching as the npm buld will only have to be run when the
package*.json, webpack config, or the translation files change.

Switched also to using python:3-slim-buster for the built image,
so there's no need to install python and pip through apt.

* dockerfile: 1-char change to appease the linter

This should have no practical effect except to silence the warning from
the Dockerfile linter and remove the problem reported by it in VS Code.

Strictly, the multi-file form of COPY should have a destination ending
in a path separator ( / or \ ), although practically . is a directory
(the root in this case).

* docker: copy over generated manifest.json

* docker: don't mask throat dir with bind mount

This commit changes what is mounted into the container with `make up` (or
`docker-compose up`).

Previously, the entire repo was mounted at /throat, but this has the
unfortunate side effect of masking any file created or altered during the
Docker build process that is copied into /throat with the version that exists
on the host (or nothing).

This is the reason, for example, that it was necessary to ensure that
`app/manifest.json` was present on the host before attempting to bring up the
application. It also affects the other files built by webpack and the
translations when they are built via Docker.

This change mounts only app/html and app/templates so these can still be edited
and changes will be picked up by the application as before.

This commit also removes the instruction in the readme to generate the
manifest.json file before developing via Docker, as this is now not necessary
as manifest.json is created when Docker runs `npm run build`.