models: reflect de-facto constraints from app code
Quite a few model fields were declared as being nullable, but various parts of
the application throw exception if these fields are actually null.

In some cases I've included defaults for these fields based on the arguments
used to construct the models elsewhere in the codebase.

There could well be other nullable fields that are problematic, but I've only
changed those that came up in the course of writing tests for one of the admin
routes.