Tests: Cache the Flask app
This change changes the `app` fixture to only create the Flask
application object once per unique configuration dictionary.

It's important to note that this means that tests share a Flask
application, though it appears that this is safe to do as the app
context is pushed before each test and popped after each test.

In return, this makes the tests run substantially faster.

Caching is done through the use of lru_cache from the standard library.
Since lru_cache can only cache hashable objects, the pyrsistent library
and a custom decorator are used to 'freeze' the dict object into a
hashable type, and then this is 'thawed' inside the get_app function
itself, so create_app receives a normal dict as expected.