Admin: Delete announcement only via POST
To remove a potential (but unlikely) CSRF vulnerability, restrict
deleting the current announcement to POST requests only, therefore
requiring a valid CSRF token.

This change includes changes to the delete view itself, and changes to
the admin index view and template to change the link over to a real
form.