Squashed commit for all CSRF tests
This commit squashes together a lot of work on tests for several
destructive routes that were previously unprotected against CSRF attacks
(plus the 'make announcement' route, which was fine, in order to better
understand the 'delete announcement' route).

Included are new tests for:

- Making and deleting announcements,
- Toggling registration/posting/captchas,
- User creation of invite codes.

In addition, there are new test helpers for creating models, logging-in
users, retrieving CSRF tokens (without making a request) etc.

- - -

I would rather have not squashed everything together but it was
unavoidable as the original work was interleaved, between
characterisation tests for the GET-accessible routes, and then TDD
implementation of the POST-only routes. And then this cycle was repeated
for the user 'create invite' route.

This resulted in a changeset that was really too large and was split
apart into its component parts. Because the tests changed significantly
over the characterisation-TDD cycles, I couldn’t come up with a set of
commits that would make sense when applied the other related pull
requests, whose changes they depend on.

So the most sensible thing was to commit only the end result, though
please be assured that this new suite of tests was definitely *arrived
at* through a process of understanding the previous behaviour before
refactoring to the new (CSRF-protected) behaviour and the matching
tests.

The original history of the whole set of changes will be kept available
at robjwells/throat/close-csrf-hole-for-do-actions in case it's useful
for fixing problems or for any other interest.